scale: Use int for translation target
In case the scale = 1.0, if the position isn't clamped to int precision,
the view appears blurry. This helps keep it sharp.

+++ plugins/single_plugins/scale.cpp
@@ -696,8 +696,8 @@ class wayfire_scale : public wf::plugin_interface_t
 
                 double scale_x = width / vg.width;
                 double scale_y = height / vg.height;
-                double translation_x = x - vg.x + ((width - vg.width) / 2.0);
-                double translation_y = y - vg.y + ((height - vg.height) / 2.0);
+                int translation_x = x - vg.x + ((width - vg.width) / 2.0);
+                int translation_y = y - vg.y + ((height - vg.height) / 2.0);
 
                 scale_x = scale_y = std::min(scale_x, scale_y);
                 if (!allow_scale_zoom)
